feat(core): validate aliases in overrides
