+++ docs/guides/4-custom-rulesets.md
@@ -327,7 +327,7 @@ For now the JSON, YAML, and JS, are all being maintained, and there are no curre
 Targeting certain parts of an OpenAPI spec is powerful but it can become cumbersome to write and repeat complex JSONPath expressions across various rules.
 Define aliases for commonly used JSONPath expressions on a global level which can then be reused across the ruleset.
 
-Aliases can be defined in an array of key-value pairs at the root level of the ruleset.
+Aliases can be defined in an array of key-value pairs at the root level of the ruleset, or alternatively, within an override.
 It's similar to `given`, with the notable difference being the possibility to distinguish between different formats.
 
 **Example**
@@ -395,6 +395,48 @@ aliases:
 ```
 
 Rulesets can then reference aliases in the [given](#given) keyword, either in full: `"given": "#Paths"`, or use it as a prefix for further JSONPath syntax, like dot notation: `"given": "#ParameterObject.name"`.
+
+Bear in mind that an alias has to be explicitly defined in either at the top-level or inside an override.
+This is to avoid ambiguity.
+
+```yaml
+aliases:
+  Stoplight:
+    - "$..stoplight"
+overrides:
+  - files:
+      - "*.yaml"
+    rules:
+      value-matches-stoplight:
+        message: Value must contain Stoplight
+        given: "#Stoplight" # valid because declared at the root
+        severity: error
+        then:
+          field: description
+          function: pattern
+          functionOptions:
+            match: Stoplight
+  - files:
+      - "**/*.json"
+    aliases:
+      Value:
+        - "$..value"
+    rules:
+      truthy-stoplight-property:
+        message: Value must contain Stoplight
+        given: "#Value" # valid because declared within the override block
+        severity: error
+        then:
+          function: truthy
+  - files:
+      - legacy/**/*.json
+    rules:
+      falsy-value:
+        given: "#Value" # invalid because undeclared both at the top-level and the override. Note that this could be technically resolvable for some JSON documents, because the previous override block has the alias, but to spare some headaches, we demand an alias to be explicitly defined.
+        severity: error
+        then:
+          function: falsy
+```
 
 > This will be followed by our core rulesets providing a common set of aliases for OpenAPI and AsyncAPI so that our users don't have to do the work at all. If you have ideas about what kind of aliases could be useful leave your thoughts [here](https://roadmap.stoplight.io).
