+++ packages/core/src/ruleset/__tests__/__fixtures__/overrides/aliases/scope.ts
@@ -44,6 +44,9 @@ const ruleset: RulesetDefinition = {
     },
     {
       files: ['legacy/**/*.json'],
+      aliases: {
+        Value: ['$..value'],
+      },
       rules: {
         'falsy-value': {
           given: '#Value',
